moved barycentric second into the correct file (was in newton_test.py), adjusted the imports to the new file system
