+++ internal_project/tests/unit_tests/automatic_tests/pipeline_entities/pipeline/default_interpolation_cores/aitken_neville_test.py
@@ -1,7 +1,7 @@
 import jax.numpy as jnp
 import unittest
 from test_interpolation_core_base import TestInterpolationCoreBase
-from interpolants.default_interpolants.aitken_neville_interpolant import AitkenNevilleInterpolant
+from functions.defaults.default_interpolants.aitken_neville_interpolant import AitkenNevilleInterpolant
 from pipeline_entities.pipeline.component_entities.default_components.default_interpolation_cores.aitken_neville_interpolation_core import (
     AitkenNevilleInterpolationCore,
 )
